refactor: outputView
